Add files via upload
Just a note that you will have to update the file location for the Database file connection from my download location to the file location on your computer, so it won't be c:/Users/yossa/Downloads/Valid Temp Range Final Project.db but whatever the file location is on your computer. I have successfully connected to the DB file and have started working with making the bar graph and the different SQLite commands.